refactor: styledcarousel for custom
