+++ components/carousel/FeaturedCarousel.tsx
@@ -1,12 +1,14 @@
-import { Carousel, CarouselProps } from "./Carousel"
+import { StyledCarousel, StyledCarouselProps } from "./StyledCarousel"
 import { SwipeCarousel } from "@/components/carousel/SwipeCarousel"
 import React from "react"
 
-export const FeaturedCarousel: React.FC<CarouselProps> = ({ carouselData }) => {
+export const FeaturedCarousel: React.FC<StyledCarouselProps> = ({
+  carouselData,
+}) => {
   return (
     <>
       <div className="hidden lg:block">
-        <Carousel carouselData={carouselData} />
+        <StyledCarousel carouselData={carouselData} />
       </div>
       <SwipeCarousel className="lg:hidden" carouselData={carouselData} />
     </>
