+++ components/carousel/SwipeCarousel.tsx
@@ -7,9 +7,9 @@ import RenderIcon from "@/components/ui/RenderIcon"
 import { Badge } from "@/components/ui/Badge"
 import {
   Dots,
-  CarouselItem,
+  StyledCarouselItem,
   Organization,
-} from "@/components/carousel/Carousel"
+} from "@/components/carousel/StyledCarousel"
 import {
   StyledDialog,
   StyledDialogContent,
@@ -35,7 +35,7 @@ const SPRING_OPTIONS: Transition = {
 }
 
 interface CarouselProps {
-  carouselData: CarouselItem[]
+  carouselData: StyledCarouselItem[]
   className?: string
 }
 
@@ -72,7 +72,7 @@ export const SwipeCarousel = ({ carouselData, className }: CarouselProps) => {
         onDragEnd={onDragEnd}
         className="flex cursor-grab items-center active:cursor-grabbing"
       >
-        <Cards carouselItems={carouselData} cardIndex={cardIndex} />
+        <Cards StyledCarouselItems={carouselData} cardIndex={cardIndex} />
       </motion.div>
 
       <Dots
@@ -85,14 +85,14 @@ export const SwipeCarousel = ({ carouselData, className }: CarouselProps) => {
 }
 
 interface CardsProps {
-  carouselItems: CarouselItem[]
+  StyledCarouselItems: StyledCarouselItem[]
   cardIndex: number
 }
 
-const Cards = ({ carouselItems, cardIndex }: CardsProps) => {
+const Cards = ({ StyledCarouselItems, cardIndex }: CardsProps) => {
   return (
     <>
-      {carouselItems.map((item, idx) => {
+      {StyledCarouselItems.map((item, idx) => {
         return (
           <motion.div
             key={idx}
@@ -165,7 +165,7 @@ const StyledDialogCard = ({
   attribution,
   buttons,
   organizations,
-}: CarouselItem) => {
+}: StyledCarouselItem) => {
   return (
     <StyledDialogContent className="max-h-3xl flex max-h-[95vh] w-[95vw] flex-col items-center overflow-y-auto rounded-xl border-none bg-white p-0 text-slate-600 sm:w-[90vw] md:w-[90vw] lg:max-w-3xl">
       <Image
